linebot/requirements.txt
flask
line-bot-sdk
google-generativeai
pyngrok
psutil
requests